Rename wasm32-wasi-preview2 to wasm32-wasip2
