+++ compiler/rustc_span/src/symbol.rs
@@ -1881,9 +1881,9 @@ symbols! {
         vtable_align,
         vtable_size,
         warn,
+        wasip2,
         wasm_abi,
         wasm_import_module,
-        wasm_preview2,
         wasm_target_feature,
         while_let,
         windows,
