+++ src/doc/rustc/src/platform-support/wasm32-wasip2.md
@@ -1,9 +1,9 @@
-# `wasm32-wasi-preview2`
+# `wasm32-wasip2`
 
 **Tier: 3**
 
-The `wasm32-wasi-preview2` target is a new and still (as of January 2024) an
-experimental target. This target is an extension to `wasm32-wasi-preview1` target,
+The `wasm32-wasip2` target is a new and still (as of January 2024) an
+experimental target. This target is an extension to `wasm32-wasip1` target,
 originally known as `wasm32-wasi`. It is the next evolution in the development of
 wasi (the [WebAssembly System Interface](https://wasi.dev)) that uses the WebAssembly
 [component model] to allow for a standardized set of syscalls that are intended to empower
