+++ tests/assembly/targets/targets-elf.rs
@@ -489,7 +489,7 @@
 //@ revisions: wasm32_wasi_preview1_threads
 //@ [wasm32_wasi_preview1_threads] compile-flags: --target wasm32-wasi-preview1-threads
 //@ [wasm32_wasi_preview1_threads] needs-llvm-components: webassembly
-//@ revisions: wasm32_wasip1
+//@ revisions: wasm32_wasip2
 //@ [wasm32_wasip2] compile-flags: --target wasm32-wasip2
 //@ [wasm32_wasip2] needs-llvm-components: webassembly
 //@ revisions: wasm64_unknown_unknown
